pyproject.toml
[project]
name = "wjecdocumentscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "google-genai>=1.49.0",
    "json-repair>=0.53.0",
    "language-tool-python>=2.9.5",
    "marker-pdf",
    "markitdown[pdf]>=0.1.3",
    "pyspellchecker>=0.8.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "pystache>=0.5.0",
    "pydantic>=2.12.4",
    "mistralai>=1.9.11",
    "isort>=7.0.0",
]

[tool.uv.sources]
marker-pdf = { git = "https://github.com/datalab-to/marker.git" }

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    # Ruff is the fast linter + autofixer; we install it as a dev tool
    "ruff>=0.13.0",
]

[tool.ruff]
# Keep formatting similar to Black defaults used across the project
line-length = 200
# Exclude large output or virtualenv dirs from scanning
exclude = [".venv", "build", "dist", "Documents", "data", "badges", "scripts", "tests"]
extend-select = ["E", "F", "I"]
